pyproject.toml
[project]
name = "votabien_pe"
version = "0.1.0"
description = ""
authors = [
    {name = "Anthony Villazana Yllesca"}
]
license = {text = "Propietary"}
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi (>=0.118.0,<0.119.0)",
    "pydantic-settings (>=2.11.0,<3.0.0)",
    "uvicorn[standard] (>=0.37.0,<0.38.0)",
    "sqlalchemy (>=2.0.43,<3.0.0)",
    "psycopg2-binary (>=2.9.10,<3.0.0)",
    "sqlmodel (>=0.0.25,<0.0.26)",
    "cuid2 (>=2.0.1,<3.0.0)",
    "resend (>=2.15.0,<3.0.0)",
    "argon2-cffi (>=25.1.0,<26.0.0)",
    "python-json-logger (>=3.3.0,<4.0.0)",
    "pyjwt (>=2.10.1,<3.0.0)",
    "email-validator (>=2.3.0,<3.0.0)",
    "python-multipart (>=0.0.20,<0.0.21)",
    "alembic (>=1.17.0,<2.0.0)"
]

[tool.poetry]
package-mode = false

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

